+++ DepRegAttributes.ExampleLibrary/ServiceCollectionExtentions.cs
@@ -1,5 +1,7 @@
 ﻿using Microsoft.Extensions.DependencyInjection;
 
+using DepRegAttributes.ExampleLibrary.DepRegAttributes;
+
 namespace DepRegAttributes.ExampleLibrary
 {
     //You need to write a service collection extention in the